[batch] Restructure halt to improve timing characteristics
Rather than build long timeouts into the stopping of each target, the
halting proceeds in phases. First all scripts in the milk category are
killed for each target, then build, then hack, grow and weaken.

In between each phase once all those scripts have been killed we wait
for a long period of time to give the scripts a chance to actually end
and some time for the follower scripts in that batch to also complete.