[get-contracts] Don't wait for a write to the contractPort
If the contract solver script errors for some reason, then this causes
the script to hang. We don't want to force the solver scripts to
respond on the port in case of an error, so it's better to wait until
the script dies and then read from the port even if it has the
sentinel empty port value.