[batch] Move hostMessage type id to recv loop
No need to have that in Sodium.

I think the rule is going to be not having any usages of `ns` APIs
inside Sodium. This should be helpful for both the purity of the
Sodium functions and hopefully keep us from accidentally calling any
`ns` APIs concurrently with a call to `ns.sleep`.