[batch] Rewrite manage script to not use Sodium
I don't think it offers any real advantages and using it requires some
gymnastics that add more complexity for no benefit.

Too much of what needs to be done requires gathering data from `ns`
calls, and making any changes also requires side-effects of `ns`
APIs. But we have to periodically run `ns.sleep` to not freeze the UI,
and if these `ns` APIs are used in the listeners to Sodium `Streams`
then there is some overlap that causes an error.