[ports] Send workers and targets to different ports
The idea here is that the management script doesn't actually need to
know about the worker hosts because it's not going to manage the
memory. Instead, it's going to send a message to the memory manager
requesting a chunk of memory for whatever task it needs, and the
memory manager will allocate chunks of memory from the available hosts
and then send that information back to the manager.

So all targets get sent to the batch manager, and the workers get sent
to the memory manager.