[refactor] Move MEMORY_PORT from /util/ports to /batch/client/memory
It conceptually makes more sense for it to be there and it reduces the
number of dependencies for a few scripts.

Probably an even better refactoring would be to make sending a worker
message a single operation on the MemoryClient. Then it could make
sense to have a "removeWorker" or "updateWorker" operation as well for
when updating personal servers.