[batch] Allow giving unique names to allocation release handlers
There was a subtle memory leak in the sow script because it made two
allocations but registered both `atExit` handlers using the default
name, thus one is overriden by the other and one of them leaks. Now
they should both be freed. Though perhaps it should be more automatic
than needing the end user to choose a unique name?