[batch] Don't reorder batch phases
Ordering them by start time seems reasonable on the surface, but it's
better to start theme in the order that we want them to finish. Any
implicit delays from each Netscript action taking a minimum amount of
time then get stacked in the same order that we want the scripts to
end, and the last script should be final one to end letting us check
whether that script is still running to determine if that batch is
finished or not.