[batch] Make TransferableAllocation.releaseAtExit more correct
Previously, I (or maybe Codex?) incorrectly used the same
`registerAllocationOwnership` function that is used by processes
registering an allocation claim by just the id of the allocation.

However, when I added the ability to claim only partial ownership of
an allocation, this includes a bunch of information about the current
process because that's the major use-case for claiming an allocation.

However, if the process is using the `TransferableAllocation` class,
then they actually still have the information about the entire
allocation and so we can and should register it to send a release for
the entire allocation.

This bug may have been messing with the allocation of harvest scripts
because they request a large block of memory that they then hold onto.