[batch] Wait for the batch ending period to time batches out correctly
I was calculating the time to sleep incorrectly under the subconcious
assumption that I was spawning the batches at
`CONFIG.batchInterval`. But I actually spawn each batch with the extra
timer so only waiting one batchInterval was causing the batch endings
to interleave in exactly the way we don't want.