[services] Schedule initial walk immediately
Previously the Discovery service set `lastWalk = 0` meaning that it
would wait the `walkRate` to walk the network looking for hosts.

This is a useful way to test that the subscriptions work, but it's not
what we want since it means at startup the other services won't
receive any hosts when they request the current list.