[util] WIP: add a new Config generating class
The idea is that we can give it a prefix, and a list of keys and their
default values and the class will create a property on the class for
each key in the list with dynamically created getter and setter
functions.

It's not working currently because of some type error that I don't
fully understand.