[batch|services|stock] Avoid data race in port reading by calling nextWrite before reading
Previously, when calling `NetscriptPort.nextWrite` after reading the
port until it was empty left an opportunity for a data race if another
port wrote to the port between when we finish reading from the port
and when we call `nextWrite`. I'm not sure that this is possible in
Netscript since I don't think there is true multi-processing happening
in the Bitburner runtime.

This brings up another opportunity for a data race actually where we
might set the `messageWaiting` flag to false, call `nextWrite` and
then have a new message pushed to the port. We would then read all
available messages, including the message that triggered `nextWrite`,
thus meaning that the port is actually empty, and yet
`messagesWaiting` is true. This is a much more benign data race though
because instead of ignoring a waiting message until yet another one is
written to the port we'll just enter the if statement on the next main
loop iteration, but the for loop will complete immediately.