[batch] Fix sow task cycling with task selector
Had a weird bug with this one where the Task Selector was correctly
identifying that the target needed to be grown a small
amount (<5%). The binary search having an initial low bound of zero
was causing the calculated threads to resolve to zero thus causing the
sow task script to quit saying "all done!" without doing
anything. Repeat ad nauseam.