[batch] Use consistent check for launching fallback limited RAM tasks
We use a strict `freeRam > x` check because trying to launch a task
that uses the exact amount of free RAM is almost guaranteed to
fail. It's an extremely marginal case, and checking free RAM still
leaves us open to trying to launch a task into a RAM environment that
contains only fragments to small to allocate the task.