[services] Add a new memory tag flag API for tagging helper scripts with allocation ID
As part of the new strategy for more precisely handling memory
allocations, we need a way for the MemoryAllocator to identify
processes that were spawned using an allocation even if they didn't
send a claim message to the allocator. This mechanism is specifically
designed for use with short-lived processes that would generate too
much traffic to the allocator and overwhelm it's message handling.