[util] Remove unused `every` utility
We likely will never actually use this function because it effectively
caches any value passed to it. The interval value passed is very
likely to be a CONFIG value and we want to pick up updates to these
whenever possible so we can change the behavior of scripts without
restarting them.