[batch] Change `batch/w.ts` to write to port in `atExit` handler
Saw this in Darktechnomancy's example scripts and it made me
think. It's not really explained explicitly why he does it this way,
but my guess is that it avoids delaying the script ending while the
write occurs. This means less chance of desyncing between the
`harvest` script waiting for this message being woken up and the w.js
script not actually having exited and freed up it's ram yet.