[services] Reduce bootstrap RAM footprint
Using `ns.run` was actually increasing the RAM footprint because the
other bootstrap scripts use `ns.exec`, and `util/dependencies` gets
charged for `ns.exec` because it uses `RegExp.exec`.

Also, with the addition of the updater service, foodnstuff doesn't
have enough space for all basic services so we launch that on n00dles
since it fits almost perfectly.