[util] Fix parsing of boolean values from LocalStorage
I foolishly just assumed that like so many other things in Javascript,
we could easily coerce strings into booleans, specifically in this
case with the `Boolean()` constructor function. However I guess we
have to do it the old fashioned way with a plain old if statement.

Now only the string "true" returns a true value and anything else
returns false.