[batch] Handle sow task being launched with small max threads
When we reworked the logic for respecting the --max-threads argument
we implicitly assumed that we would never need fewer than the perfect
ratio of grow threads to grow a server. However, late in a Bitnode
when we have a lot of multipliers this is very possible.

The way we were calculating the sow batch logistics meant that we had
a needed number of sow batches less than 1, which seemed like
nonsense, and we had an early exit condition that was being regularly
tripped on easy targets.

Now we calculate the batch threads and totalBatches correctly
accounting for if we need fewer threads than the perfect thread ratio.