[buy-servers] Name servers with index only
Renaming the purchased servers is extremely complicated with
persistent scripts like harvest that continually respawn scripts in
their allocations.

We used to need the name of the servers to easily know how much RAM
they had but the new memory allocator display shows that directly so
we can give them generic names and just completely remove the renaming
capability.