[batch] Use totalBatches instead of overlap from SowBatchLogistics
Before this, as our grow time got faster we calculated a lower overlap
meaning more total rounds. For most of a bitnode this means we end up
calculating 2 rounds so sowing takes twice as long as it needs to.

However, in extremely late bitnodes when our hacking level is insanely
high and our grow time is extremely low we end up spawning 30-40 1
second batches that we run serially instead of a single large batch
that runs in about a second. This means that as our grow time went
down our actual grow time to needed grow time was growing worse and
worse.

The cause of this idea was trying to run sow batches in a serial
pipeline like harvest batches. But the exponential scaling of larger
sow batches far outweighs the scaling of building off of previous
batches that this idea just wasn't worth while. It is helpful for
scaling sow tasks down, but we already partially backed off from this
pipeline idea by scaling batches to the maximum size we have for our
current host when spawning. This is the final piece in backing off of
the idea of pipelining grow tasks.