[corp] Combine intervalId and eating state into a component-local mutable ref
The module level variable was preventing restarting the script to
restart eating once the interval stopped running, since subsequent
invocations of the script would see the retained module level state of
`eating` being true.

This new structure should mean that closing the log and running a new
instance of the eat script is enough to reset it if the script gets
into a bad state.