[bladeburner] Only average spread when it is non-zero
As our skills grow some easier contracts and operations will have a
high enough success chance that even when there is uncertainty from a
bad population estimate, the spread will still be zero. In this case
it is better to estimate uncertainty only by looking at contracts and
operations with a non-zero spread.