[util] Remove dangerous waitForCompletion option
There is really no valid use-case for it that I can think of. If you
don't want your script to wait for the command to complete then don't
await the promise. However, in order to correctly serialize commands
to the terminal we should always wait for timed commands to complete
internally.