[services] Remove service daemon tests meant to be run inside Bitburner
Now that all protocols are being ported to use the new protocol
abstraction, these test are unnecessary. The protocol abstraction has
a robust set of unit and integration tests that run as part of the
Jest test suite that provide much more confidence that the abstraction
is working correctly. The abstraction is also type-safe enough that
typescript should throw an error if the protocol is being used
incorrectly from typescript code, and there are runtime checks for if
typescript errors are overridden with bad casts or the protocol is
used from Javascript code.