[services] Fix GrowableAllocation releaseAtExit to not call ns fns during atExit callback
Because the script has already been killed by the time the atExit
callback runs, calling `ns` functions throws an error so we need to
capture the our self-process description earlier so we don't need to
call `ns.self()` during the `atExit` handler.

This is a subtle memory leak and was definitely introduced quite a
while ago. I wonder if this is the source of our leaking memory?