[util] Warn on timeout but keep waiting
On paper, having timeouts seems like a solid idea. In practice my
scripts and my messaging architecture was built with guaranteed
delivery and receipt of messages in mind. If one script times out and
throws an error then this is likely to cause cascading failures.

Maybe having timeouts is the right thing to do, but currently, adding
timeouts would basically mean that we need to add a ton of try-catch
statements and retries and other nonsense to a lot of my scripts.

The reality is, these ports are an in-memory data structure and it's
no more fallible to write to and read from a port than it is to push
and pop from an array. Timeouts are a technique that is useful for
real, fallible networks where the server on the other end might not be
listening, or the network might fail before you get your response, or
all the sys admins in the world might accidentally take their lunch
break at the same time and the apocalypse could start. Who knows?!?!

But in Bitburner, we don't need to worry about that.