[services] Call zero RAM Netscript functions in the client process
They don't cost anything so they can't cause the dynamic RAM usage to
exceed static calculations so there is no need to pay for the
round-trip to the server.

It's a bit pointless to call zero-cost RAM functions but this is
probably a better choice than just rejecting zero cost functions with
a thrown exception.