[expoloits] Change dev-menu to be js instead of ts
There are no types whatsoever so I would just need to add a bunch of
lines disabling or suppressing specific errors to no benefit.

This file is some janky monkey patching that works and I don't want to
mess with it to try and make it "safer". If it fails, that's fine and
I'll deal with it if/when that happens.