[lib] Stricter types in sodium/Operational
This is tricky. There is a lot of code that plays fast and loose with
nullable values, and a lot of it is very not obvious whether it is
correct or not. A lot of these changes I've been able to convince
myself of the correctness after looking at the code a little bit, but
a lot of it is substantially harder to verify.

I probably should have added explanatory comments to all of these with
suggestions as to why the code is correct as written.

One thing I've noticed that is probably an actual improvement prompted
by the stricter types is that the code seems largerly unconcerned with
possible undefined values, typically only checking for specifically `x
=== null` instead of the "nullish check" of `x == null`. I don't think
any of the code is specifically treating `undefined` differently than
null in a semantically important way, but it's again, quite hard to
tell.

It's quite difficult and I'm not really sure it's worthwhile to make
Sodium compile with strict null checking on.