[automation] Handle buying very high levels of NFG more robustly
When the needed reputation or the donation for reputation exceed the
max safe integer we can possibly get rounding that makes it so that we
fail to buy the neuroflux governor because we don't have enough
reputation even after donating to get the reputation. Donating again
usually fixes this issue so we just loop the donate purchase process
until purchasing succeeds.